move downloadHandler to core components folder
